requirements.txt
yacs
flask
torch
flask_cors
opencv-python
faiss-cpu
lshashpy3
torchvision
tqdm
scikit-learn